fix(dsv4): int64 program_id in _fused_rmsnorm_rope_kernel (64K-prompt overflow)
Same pitfall as v4_rmsnorm (commit a6f720e85): the prefill Q site
flattens [B, S, H, D] to [B*S*H, D] before launching one program per
row, and ``pid * x_stride_n`` (default int32) silently overflows when
N*D >= 2^31.  V4-Flash 64K warmup hits this at B=1 S=65600 H=64 D=512:
  max pid * x_stride_n = 4198399 * 512 = 2.15B > INT32_MAX
Pointer wraps; CUDA reports the deferred error at the next sync site.

Cast pid to int64 once at the top of the kernel — this also widens
``freq_idx = pid // freq_stride_n`` automatically.  Bit-exact, no perf
cost.  Verified locally on v4_flash_64k_tp1_sm100 with
CUDA_LAUNCH_BLOCKING=1: prefill completes and produces a coherent
50-token decode (no more ILLEGAL_ADDRESS at the rope kernel).

+++ rtp_llm/models_py/modules/dsv4/_fused_rmsnorm_rope_triton.py
@@ -37,7 +37,12 @@ def _fused_rmsnorm_rope_kernel(
     HAS_WEIGHT: tl.constexpr,
     BLOCK_D: tl.constexpr,
 ):
-    pid = tl.program_id(0)
+    # Cast pid to int64 BEFORE the row-stride multiply: the prefill Q
+    # site flattens [B, S, H, D] to [B*S*H, D], so N can exceed 2^22 and
+    # ``pid * x_stride_n`` silently overflows int32 (e.g. S=65600 H=64
+    # D=512: max pid * stride = 4198399 * 512 = 2.15B > INT32_MAX →
+    # CUDA_ERROR_ILLEGAL_ADDRESS).  Same pitfall as v4_rmsnorm.
+    pid = tl.program_id(0).to(tl.int64)
     # Tokens per freq slot — decouples the kernel from decode/prefill layout:
     #   decode Q  [B, 1, H, D] + freqs [B, RD/2]     -> freq_stride_n = H
     #   decode KV [B, 1,    D] + freqs [B, RD/2]     -> freq_stride_n = 1
